Write a function that returns the sum of all its parameters.
    Prototype: int sum_them_all(const unsigned int n, ...);
    If n == 0, return 0

+++ variadic_functions/0-sum_them_all.c
@@ -10,13 +10,13 @@ int sum_them_all(const unsigned int n, ...)
 {
 	va_list args;
 	unsigned int i;
-	unsigned int sum = 0;
+	double sum = 0;
 
 	if (n == 0)
 		return (0);
 	va_start(args, n);
 	for (i = 0; i < n; i++)
-		sum += va_arg(args, unsigned int);
+		sum += va_arg(args, double);
 	va_end(args);
 	return (sum);
 }
